[ANDROSDK-988] Move D2 and D2Manager to same package and remove builder
