+++ core/src/androidTest/java/org/hisp/dhis/android/core/D2ManagerRealIntegrationShould.java
@@ -26,11 +26,13 @@
  * SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  */
 
-package org.hisp.dhis.android.core.d2manager;
+package org.hisp.dhis.android.core;
 
 import com.facebook.stetho.okhttp3.StethoInterceptor;
 import com.google.common.collect.Lists;
 
+import org.hisp.dhis.android.core.D2Configuration;
+import org.hisp.dhis.android.core.D2Manager;
 import org.hisp.dhis.android.core.common.ObjectWithUid;
 import org.hisp.dhis.android.core.data.server.RealServerMother;
 import org.hisp.dhis.android.core.user.UserCredentials;
