+++ core/src/androidTest/java/org/hisp/dhis/android/core/TeisCallRealIntegrationShould.java
@@ -31,7 +31,6 @@ package org.hisp.dhis.android.core;
 import android.util.Log;
 
 import org.hisp.dhis.android.core.arch.call.D2Progress;
-import org.hisp.dhis.android.core.d2manager.D2Factory;
 import org.hisp.dhis.android.core.trackedentity.internal.TrackedEntityInstanceStoreImpl;
 import org.hisp.dhis.android.core.utils.integration.real.BaseRealIntegrationTest;
 import org.junit.Before;
