+++ core/src/main/java/org/hisp/dhis/android/core/D2ConfigurationValidator.java
@@ -26,7 +26,7 @@
  * SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  */
 
-package org.hisp.dhis.android.core.d2manager;
+package org.hisp.dhis.android.core;
 
 import android.content.Context;
 import android.content.pm.ApplicationInfo;
