+++ core/src/main/java/org/hisp/dhis/android/core/TLSSocketFactory.java
@@ -26,7 +26,7 @@
  * SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  */
 
-package org.hisp.dhis.android.core.d2manager;
+package org.hisp.dhis.android.core;
 
 import java.io.IOException;
 import java.net.InetAddress;
